requirements.txt
Flask
selenium
geopy
requests
pandas
beautifulsoup4
python-dotenv
